cv-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic-settings==2.2.1
Pillow==10.3.0
python-multipart==0.0.9
torch==2.3.0
transformers==4.41.0
